refactor: fix useState hook usage in search HOC to prevent React context errors
