+++ resources/ts/components/DegreeProgramCombinations/hoc/withSearchedDegreeProgramPosts.tsx
@@ -1,5 +1,6 @@
-import { useState } from 'react';
+import React, { useState } from 'react';
 
+import { createHigherOrderComponent } from '@wordpress/compose';
 import type { Post } from '@wordpress/core-data';
 import { store as coreStore } from '@wordpress/core-data';
 import { withSelect } from '@wordpress/data';
@@ -8,13 +9,13 @@ import { store as editorStore } from '@wordpress/editor';
 import serverData from '../../../util/serverData';
 
 import { EntitySelectorProps } from '../../../defs';
+import { SelectorProps } from '../../TermSelector/defs';
 
-const withSearchedDegreeProgramPosts = withSelect(
+const withSearchedPostsSelect = withSelect(
 	(
 		select,
 		ownProps: EntitySelectorProps< Post< 'view' > >
 	): Partial< EntitySelectorProps< Post< 'view' > > > => {
-		const [ search, setSearch ] = useState( '' );
 		const { getCurrentPost } = select( editorStore.name );
 		const { getEntityRecords } = select( coreStore.name );
 
@@ -27,16 +28,34 @@ const withSearchedDegreeProgramPosts = withSelect(
 				order: 'asc',
 				_fields: 'id,degree_program',
 				context: 'view',
+				search: ownProps.search,
 			},
-			search,
 			...( currentPost ? { exclude: currentPost.id } : {} ),
 		} );
 
 		return {
 			searchedEntities: posts ?? [],
-			setSearch,
 		};
 	}
+);
+
+const withSearchedDegreeProgramPosts = createHigherOrderComponent(
+	( WrappedComponent: React.FC< Partial< EntitySelectorProps > > ) =>
+		( props: SelectorProps ) => {
+			const [ search, setSearch ] = useState( '' );
+			const EnhancedComponent: React.FC<
+				Partial< EntitySelectorProps >
+			> = withSearchedPostsSelect( WrappedComponent );
+
+			return (
+				<EnhancedComponent
+					{ ...props }
+					search={ search }
+					setSearch={ setSearch }
+				/>
+			);
+		},
+	'withSearchedDegreeProgramPosts'
 );
 
 export default withSearchedDegreeProgramPosts;
